% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCaps.R
\name{mapFETISH}
\alias{mapFETISH}
\title{Map the paired-end FETISH data.}
\usage{
mapFETISH(index, R1, R2, output, nthreads, logfile = NULL, ...)
}
\arguments{
\item{index}{character string giving the basename of the Subread index file.}

\item{R1}{forward read (_R1) fastq, output of \code{\link{trimFlash}} command.}

\item{R2}{reverse read (_R2) fastq, output of \code{\link{trimFlash}} command.}

\item{output}{output file name (without ".bam" extention)}

\item{nthreads}{number of threads to use for mapping.}

\item{logfile}{a log file to write the processing message.}

\item{...}{additional arguments passed to the RSubread::align function.}
}
\value{
bamfile A mapped BAM file for the sample.
}
\description{
Map the paired-end FETISH data.
}
\examples{
r1 <- system.file("extdata", "testout_R1.fastq.gz", package = "FATSCapR")
r2 <- system.file("extdata", "testout_R2.fastq.gz", package = "FATSCapR")
\dontrun{
mapFETISH(index,R1 = r1, R2 = r2, output = "test_mapped", nthreads = 10, logfile=NULL)
}

}

